refactor(swap): address #4331 review feedback (Issues 2-5 + @MainActor)
Five review-feedback items from #4331:

- **@MainActor on SwapTransaction (CodeRabbit)**: the class is an
  ObservableObject with @Published vars used throughout SwiftUI views.
  Marked @MainActor; cascade required @MainActor on the swap-specific
  methods in BlockChainService, SecurityScannerService /
  SecurityScannerServiceProtocol / SecurityScannerTransactionFactory /
  SecurityScannerViewModel, and on SwapCryptoLogic. Reference-identity
  ==/hash(into:) on SwapTransaction marked nonisolated so SwapRoute's
  derived Equatable can be invoked off-MainActor.

- **Issue 2 — Double-trigger race on load**: SwapDetailsScreen called
  `detailsViewModel.load(...)` then `tx.fromCoin = fromCoin`. `load()`
  already sets tx.fromCoin; the second assignment re-fired
  `onChange(of: tx.fromCoin)` → redundant racing quote fetch. Removed
  the redundant assignment.

- **Issue 3 — Fragile hard-coded pop count**: SwapKeysignScreen used
  `router.navPath.removeLast(min(2, …))` assuming verify was always
  exactly 2 levels back. Replaced with `router.navigateBack(matching:)`
  scanning history for the `.verify` route — robust to deep-links that
  add routes before `.root`.

- **Issue 4 — SwapRouteBuilder redundant indirection**: SwapRouter just
  instantiated a SwapRouteBuilder() and called one of five methods. The
  five @ViewBuilder methods now live directly on SwapRouter, matching
  the pattern used by Home/Vault route flows. SwapRouteBuilder.swift
  deleted; HomeRouteBuilder and VaultRouteBuilder updated to construct
  SwapRouter() instead.

- **Issue 5 — Timer + refreshData duplicated across two screens**:
  extracted `SwapRefreshTickModifier` + `.swapRefreshTick { … }` view
  modifier. SwapDetailsScreen and SwapVerifyScreen both use it now.

Issue 1 (@Model objects in route associated values) ships as a separate
commit due to the larger structural surface.
