+++ VultisigApp/VultisigApp/Features/Swap/ViewModels/SwapCryptoLogic.swift
@@ -10,6 +10,7 @@ import Mediator
 import SwiftUI
 import WalletCore
 
+@MainActor
 enum SwapCryptoLogic {
     private static let swapService = SwapService.shared
     private static let blockchainService = BlockChainService.shared
